Elongwheat with runing test and example. In this version, tillering is implemented as a hack (copy of the corresponding leaves on the MS).
From this commit, the branch will hold new features under development which will be then merged with the trunk.
The versions on this branch should be stable (running test and examples). Please start new branches (features) from the develop branch to develop new features for the upcoming or a distant future release. Once stable, feature branches should be merged with develop branch only.